(fix) 방나가기 버튼 옮김
